test(s2n-quic): refactor integration tests into submodules
