+++ quic/s2n-quic/src/tests/blackhole.rs
@@ -1,0 +1,48 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use super::*;
+
+fn blackhole(model: Model, blackhole_duration: Duration) {
+    test(model.clone(), |handle| {
+        spawn(async move {
+            // switch back and forth between blackhole and not
+            loop {
+                delay(blackhole_duration).await;
+                // drop all packets
+                model.set_drop_rate(1.0);
+
+                delay(blackhole_duration).await;
+                model.set_drop_rate(0.0);
+            }
+        });
+        client_server(handle)
+    })
+    .unwrap();
+}
+
+#[test]
+fn blackhole_success_test() {
+    let model = Model::default();
+
+    let network_delay = Duration::from_millis(1000);
+    model.set_delay(network_delay);
+
+    // setting the blackhole time to `network_delay / 2` causes the connection to
+    // succeed
+    let blackhole_duration = network_delay / 2;
+    blackhole(model, blackhole_duration);
+}
+
+#[test]
+#[should_panic]
+fn blackhole_failure_test() {
+    let model = Model::default();
+
+    let network_delay = Duration::from_millis(1000);
+    model.set_delay(network_delay);
+
+    // setting the blackhole time to `network_delay / 2 + 1` causes the connection to fail
+    let blackhole_duration = network_delay / 2 + Duration::from_millis(1);
+    blackhole(model, blackhole_duration);
+}
